// src/app/models/user.model.ts
export interface User{
    id: number;
    name: string;
    phoneNumber: string;
    email: string;
    position?: string;
    instagram?: string;
    twitter?: string;
    profileImageUrl?: string;
    foto: string;
    nivel: string;
    biografia: string;
    salario: number;
    fotoUrl: string;
}